% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrimMean.R
\name{getTrimMean}
\alias{getTrimMean}
\title{Compute the trimmed mean of a sample from Normal distribution}
\usage{
getTrimMean(
  n = 1000,
  N = 5000,
  mu = 0,
  sigma = 1,
  tr = 0,
  random.seed = 20230615
)
}
\arguments{
\item{n}{The size of a single simulation sampled from the Normal distribution.
An integer. Defaults to 1000.}

\item{N}{The nunber of repeated simulations. An integer. Default to 5000.}

\item{mu}{The grounding true mean of the Normal distribution. A float. Default to 0.}

\item{sigma}{The grounding true standard deviation of the Normal distribution.
A float. Default to 1.}

\item{tr}{The trim ratio, the proportion of the sample is chopped off as extreme values.
A float. Default to 0.}

\item{random.seed}{Set a random seed.}
}
\value{
A vector of trimmed sample means of size `N`. If `tr` is set as 0, it effectively
returns the untrimmed mean.
}
\description{
`getTrimMean()` computes the trimmed sample means of `N` simulations (each of size `n`)
from a Normal distribution with mean `mu` and standard deviation `sigma`. The trimmed
sample is obtained by chopping off the `tr` portion of the extreme values in a sample.
}
\examples{
# Directly get a vector of trimmed sample means of size `N`.
getTrimMean()


}
