% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{plotter}
\alias{plotter}
\title{Plotter Function}
\usage{
plotter(plot_type, n, mu, sigma, tr, N)
}
\arguments{
\item{plot_type}{takes values from ["boxplot", "MCSE", "EmpVar", "QQplot", "ECDF"]}

\item{n}{sample size of each simulation}

\item{mu}{The grounding true mean of the Normal distribution. A float. Default to 0.}

\item{sigma}{The grounding true standard deviation of the Normal distribution.
A float. Default to 1.}

\item{tr}{The trim ratio, the proportion of the sample is chopped off as extreme values.
A float. Default to 0.}

\item{N}{The nunber of repeated simulations. An integer. Default to 5000.}
}
\value{
plot
}
\description{
`plotter()` can plot the boxplot, Monte Carlo Simulation Error (MCSE), Empirical Variance
(EmpVar), QQplot, and Empirical Cummulative Distribution Function (ECDF).
}
