% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{tabler}
\alias{tabler}
\title{Tabler Function}
\usage{
tabler(n, mu, sigma, tr, N = 15000, to_kable = F, random.seed = 20230615)
}
\arguments{
\item{n}{sample size of each simulation}

\item{mu}{The grounding true mean of the Normal distribution. A float.}

\item{sigma}{The grounding true standard deviation of the Normal distribution.
A float.}

\item{tr}{The trim ratio, the proportion of the sample is chopped off as extreme values.
A float.}

\item{N}{The number of repeated simulations. An integer. Default as 15000.}

\item{random.seed}{Set a random seed.}
}
\description{
`tabler()` can output a table with the bias, Empirical Variance (EmpVar) and Monte Carlo Simulation Error (MCSE)
as the columns. The rows corresponds to the input argument (one of ['n', 'mu', 'sigma', 'tr', 'N']) that takes a
vector with length > 1.
}
\details{
Note that only one of the input argument can take a vector with length > 1, the other input arguments have to be
singular numbers. Otherwise, the function will explode.
}
\examples{
To obtain a table of bias, EmpVar and MCSE corresponding to various trim ratios:
tabler(50, 0, 1, seq(0,0.5, by=0.025))


}
