% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimFZC-package.R
\docType{package}
\name{trimFZC-package}
\alias{trimFZC-package}
\alias{trimFZC}
\title{Trim Mean Estimator of Normal Distribution}
\description{
'trimFZC' package contains 3 main functions: \code{getTrimMean()}, \code{plotter}, and \code{tabler}.
\itemize{
 \item \code{getTrimMean()}: generates trimmed mean estimates of simulated samples from Normal
 distribution, see \link{getTrimMean} for details.
 \item \code{plotter}: draws various plots for analysis of the efficiency of the trimmed
 mean estimates, see \link{plotter} for details.
 \item \code{tabler}: creates tabularized statistical summaries of the trimmed
 mean estimates, see \link{tabler} for details.
}
For demonstrative examples, see \code{vignette("vignette")}.
}
\author{
Yuling Max Chen (\email{yuling.chen@uwaterloo.ca}, developer)

Yuxuan Zhao (contributor)

Megan French (contributor)
}
